+++ common-rei/src/main/java/com/yanny/ali/compatibility/rei/ReiEntityDisplay.java
@@ -4,7 +4,8 @@ import com.yanny.ali.compatibility.common.EntityLootType;
 import me.shedaniel.rei.api.common.category.CategoryIdentifier;
 import me.shedaniel.rei.api.common.entry.EntryIngredient;
 import me.shedaniel.rei.api.common.util.EntryIngredients;
-import net.minecraft.world.entity.Entity;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.SpawnEggItem;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Unmodifiable;
@@ -12,12 +13,14 @@ import org.jetbrains.annotations.Unmodifiable;
 import java.util.List;
 
 public class ReiEntityDisplay extends ReiBaseDisplay {
-    private final Entity entity;
+    private final EntityType<?> entityType;
+    private final ResourceLocation variant;
     private final CategoryIdentifier<ReiEntityDisplay> identifier;
 
     public ReiEntityDisplay(EntityLootType entry, CategoryIdentifier<ReiEntityDisplay> identifier) {
         super(getSpawnEgg(entry), entry);
-        entity = entry.entity();
+        entityType = entry.entityType();
+        variant = entry.variant();
         this.identifier = identifier;
     }
 
@@ -26,14 +29,18 @@ public class ReiEntityDisplay extends ReiBaseDisplay {
         return identifier;
     }
 
-    public Entity getEntity() {
-        return entity;
+    public EntityType<?> getEntityType() {
+        return entityType;
+    }
+
+    public ResourceLocation getVariant() {
+        return variant;
     }
 
     @Unmodifiable
     @NotNull
     private static List<EntryIngredient> getSpawnEgg(EntityLootType entry) {
-        SpawnEggItem item = SpawnEggItem.byId(entry.entity().getType());
+        SpawnEggItem item = SpawnEggItem.byId(entry.entityType());
 
         if (item != null) {
             return List.of(EntryIngredients.of(item));