+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## []
+
+- Fixed memory leak of ClientLevel
+
 ## [1.3.5]
 
 - Fixed crash on Forge 47.4.0+